backend/requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Authentication
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3

# Pose Detection
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
Werkzeug==2.3.7
requests==2.31.0
python-multipart==0.0.6 